+++ common/service/gsmlogo.h
@@ -31,9 +31,9 @@ size_t GSM_GetBitmapSize(GSM_Bitmap * bmp);
 GSM_Error BMP2Bitmap            (unsigned char *buffer, FILE *file,GSM_Bitmap *bitmap);
 GSM_Error Bitmap2BMP            (unsigned char *buffer, FILE *file,GSM_Bitmap *bitmap);
 
-void PHONE_GetBitmapWidthHeight	(GSM_Phone_Bitmap_Types Type, int *width, int *height);
-size_t PHONE_GetBitmapSize	(GSM_Phone_Bitmap_Types Type, int width, int height);
-void PHONE_ClearBitmap		(GSM_Phone_Bitmap_Types Type, char *buffer, int width, int height);
+void PHONE_GetBitmapWidthHeight	(GSM_Phone_Bitmap_Types Type, size_t *width, size_t *height);
+size_t PHONE_GetBitmapSize	(GSM_Phone_Bitmap_Types Type, size_t width, size_t height);
+void PHONE_ClearBitmap		(GSM_Phone_Bitmap_Types Type, char *buffer, size_t width, size_t height);
 void PHONE_DecodeBitmap		(GSM_Phone_Bitmap_Types Type, char *buffer, GSM_Bitmap *Bitmap);
 void PHONE_EncodeBitmap		(GSM_Phone_Bitmap_Types Type, char *buffer, GSM_Bitmap *Bitmap);
